+++ docs/Changelog.md
@@ -5,6 +5,7 @@
 ### If upgrading from a version before 2.00 please read the notes for version 2.00 before continuing.
 
 * Fix bug where TuneIn radio stations didn't work for MPD users. (Was fixed in 2.24 but this is a better fix)
+* You can now select, upload, and modify album art right from the collection browser
 
 
 ## Version 2.22